luminous: Create the bootstrap-rbd keyring when using etcd
The static config creates a keyring for RBD_MIRROR_BOOTSTRAP_KEYRING,
which is used by rbd-mirror and also by the start script for the
mon. The etcd config was not creating this keyring, preventing mon and
rbd-mirror from starting.

Add RBD_MIRROR_BOOTSTRAP_KEYRING to the list of keyrings to generate
when using etcd.